Reset skeleton

+++ Mu.NETcms/Models/GameModels.cs
@@ -57,16 +57,16 @@ namespace Mu.NETcms.Models
             public string Name { get; set; }
             public int cLevel { get; set; }
             public int LevelUpPoint { get; set; }
-            public int Class { get; set; }
+            public Byte Class { get; set; }
             public int Experience { get; set; }
             public int Strength { get; set; }
             public int Dexterity { get; set; }
             public int Vitality { get; set; }
             public int Energy { get; set; }
-            public int MapNumber { get; set; }
-            public int MapPosX { get; set; }
-            public int MapPosY { get; set; }
-            public int CtlCode { get; set; }
+            public Int16 MapNumber { get; set; }
+            public Int16 MapPosX { get; set; }
+            public Int16 MapPosY { get; set; }
+            public Byte CtlCode { get; set; }
             public int Resets { get; set; }
             public int GrandResets { get; set; }
             public int Money { get; set; }
